+++ src/entities/enemy-types.js
@@ -12,6 +12,7 @@ import { Zombie } from './enemy-types/zombie.js';
 import { AmalgamZombie } from './enemy-types/amalgam-zombie.js';
 import { ArmoredKnight } from './enemy-types/armored-knight.js';
 import { Shounao } from './enemy-types/shounao.js';
+import { FlySwarm } from './enemy-types/fly-swarm.js';
 
 function getAnimConfig(key) {
     return ANIMATION_CONFIG[key] || {};
@@ -471,4 +472,4 @@ class ZombieDogEnemy extends CircleEnemy {
     }
 }
 
-export { BlackWolf, CircleEnemy, ZombieDogEnemy, ZombieWizard, Mutant3, SpitterZombie, FatZombie, Zombie, AmalgamZombie, ArmoredKnight, Shounao };
+export { BlackWolf, CircleEnemy, ZombieDogEnemy, ZombieWizard, Mutant3, SpitterZombie, FatZombie, Zombie, AmalgamZombie, ArmoredKnight, Shounao, FlySwarm };